ai_writing_flow/pyproject.toml
[project]
name = "ai_writing_flow"
version = "0.1.0"
description = "AI Writing Flow - Content generation with styleguide compliance"
authors = [{ name = "Vector Wave AI", email = "ai@vectorwave.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
kickoff = "ai_writing_flow.main:kickoff"
run_crew = "ai_writing_flow.main:kickoff"
plot = "ai_writing_flow.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
